Move rpcToBlockData and rpcToTxData to test utils
